requirements.txt
asgiref==3.3.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.7
pytz==2020.5
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.9
